challenge1/application/requirements.txt
google-auth==2.17.3
requests==2.31.0
Flask==2.1.0
gunicorn==20.1.0
